move generators in files
