+++ src/generators/files/index.ts
@@ -1,0 +1,8 @@
+export * from "./browser-ts.js";
+export * from "./client-ts.js";
+export * from "./main-spec-ts.js";
+export * from "./main-ts.js";
+export * from "./package-json.js";
+export * from "./server-ts.js";
+export * from "./shared-ts.js";
+export * from "./tsconfig-json.js";
