+++ src/generators/files/main-ts.ts
@@ -1,4 +1,4 @@
-import { CodeGeneratorBase } from "./code-generator-base.js";
+import { CodeGeneratorBase } from "../code-generator-base.js";
 
 export class MainTsCodeGenerator extends CodeGeneratorBase {
   public *getStatements() {
