+++ src/generators/files/shared-ts.ts
@@ -1,8 +1,8 @@
 import camelcase from "camelcase";
 import ts from "typescript";
-import * as models from "../models/index.js";
-import { toPascal } from "../utils/name.js";
-import { CodeGeneratorBase } from "./code-generator-base.js";
+import * as models from "../../models/index.js";
+import { toPascal } from "../../utils/name.js";
+import { CodeGeneratorBase } from "../code-generator-base.js";
 
 export class SharedTsCodeGenerator extends CodeGeneratorBase {
   public *getStatements() {
